Refactor code, solve floating point problems, include additional enumerations

+++ src/main/java/org/acme/albertosenserrich/provisioncalculator/constants/GeneralConstants.java
@@ -1,8 +1,8 @@
 package org.acme.albertosenserrich.provisioncalculator.constants;
 
 public final  class GeneralConstants {
-	public final static int COST_OF_RENT = 4320086;
-	public final static int DEFAULT_EXPENSE_FOR_EMPLOYE = 100;
+	public final static long COST_OF_RENT = 4320086;
+	public final static long DEFAULT_EXPENSE_FOR_EMPLOYE = 100;
 	
 
 }
